+++ apps/cowswap-frontend/src/modules/account/containers/AccountDetails/AccountIcon.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react'
+import { ReactNode, useState } from 'react'
 
 import { HoverTooltip } from '@cowprotocol/ui'
 import { useConnectionType, useWalletDetails } from '@cowprotocol/wallet'
@@ -14,9 +14,7 @@ interface AccountIconProps {
   account?: string
 }
 
-// TODO: Add proper return type annotation
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-export const AccountIcon = ({ size = 16, account }: AccountIconProps) => {
+export const AccountIcon = ({ size = 16, account }: AccountIconProps): ReactNode => {
   const walletDetails = useWalletDetails()
   const connectionType = useConnectionType()
   const [imageLoadError, setImageLoadError] = useState(false)