+++ apps/cowswap-frontend/src/modules/limitOrders/services/safeBundleFlow/index.ts
@@ -177,6 +177,7 @@ export async function safeBundleFlow({
       id: orderId,
       orderCreationHash: safeTxHash,
       kind: postOrderParams.kind,
+      quoteId: postOrderParams.quoteId,
       receiver: recipientAddressOrName,
       inputAmount,
       outputAmount,